process/services/service_a/pyproject.toml
[tool.poetry]
name = "service_a"
version = "1.0.0"
description = "Simple and fast service..."
authors = ["Fabricio Nogueira <nogsantos@gmail.com>"]

[tool.poetry.dependencies]
python = "~=3.9"
nameko = "^2.14.1"
requests = "^2.30.0"
pydantic = "^1.10.7"
python-decouple = "^3.8"
python-json-logger = "^2.0.7"
redis = "^4.5.5"
tenacity = "^8.2.2"
retry = "^0.9.2"
pytz = "^2024.1"
eventlet = "^0.39.1"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
responses = "^0.23.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | migrations
)
'''